Improve commands
