+++ cmd/root/a2a.go
@@ -6,6 +6,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/docker/cagent/pkg/a2a"
+	"github.com/docker/cagent/pkg/agentfile"
 	"github.com/docker/cagent/pkg/cli"
 	"github.com/docker/cagent/pkg/config"
 	"github.com/docker/cagent/pkg/server"
@@ -23,9 +24,9 @@ func newA2ACmd() *cobra.Command {
 	var flags a2aFlags
 
 	cmd := &cobra.Command{
-		Use:   "a2a <agent-file>",
+		Use:   "a2a <agent-file>|<registry-ref>",
 		Short: "Start an agent as an A2A (Agent-to-Agent) server",
-		Long:  `Start an A2A server that exposes the agent via the Agent-to-Agent protocol`,
+		Long:  "Start an A2A server that exposes the agent via the Agent-to-Agent protocol",
 		Example: `  cagent a2a ./agent.yaml
   cagent a2a ./team.yaml --port 8080
   cagent a2a agentcatalog/pirate --port 9000`,
@@ -47,11 +48,6 @@ func (f *a2aFlags) runA2ACommand(cmd *cobra.Command, args []string) error {
 
 	ctx := cmd.Context()
 	out := cli.NewPrinter(cmd.OutOrStdout())
-	agentFilename := args[0]
-
-	if err := setupWorkingDirectory(f.workingDir); err != nil {
-		return err
-	}
 
 	// Listen as early as possible
 	ln, err := server.Listen(ctx, fmt.Sprintf(":%d", f.port))
@@ -62,6 +58,15 @@ func (f *a2aFlags) runA2ACommand(cmd *cobra.Command, args []string) error {
 		<-ctx.Done()
 		_ = ln.Close()
 	}()
+
+	if err := setupWorkingDirectory(f.workingDir); err != nil {
+		return err
+	}
+
+	agentFilename, err := agentfile.Resolve(ctx, out, args[0])
+	if err != nil {
+		return err
+	}
 
 	return a2a.Start(ctx, out, agentFilename, f.agentName, f.runConfig, ln)
 }
