+++ cmd/root/alias.go
@@ -18,7 +18,7 @@ func newAliasCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "alias",
 		Short: "Manage aliases",
-		Long:  `Create and manage aliases for agent configurations or catalog references.`,
+		Long:  "Create and manage aliases for agent configurations or catalog references.",
 		Example: `  # Create an alias for a catalog agent
   cagent alias add code agentcatalog/notion-expert
