+++ cmd/root/feedback.go
@@ -13,7 +13,7 @@ func newFeedbackCmd() *cobra.Command {
 	return &cobra.Command{
 		Use:   "feedback",
 		Short: "Send feedback about cagent",
-		Long:  `Submit feedback or report issues with cagent`,
+		Long:  "Submit feedback or report issues with cagent",
 		Args:  cobra.NoArgs,
 		RunE:  runFeedbackCommand,
 	}
