+++ cmd/root/mcp.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/spf13/cobra"
 
+	"github.com/docker/cagent/pkg/agentfile"
 	"github.com/docker/cagent/pkg/cli"
 	"github.com/docker/cagent/pkg/config"
 	"github.com/docker/cagent/pkg/mcp"
@@ -20,9 +21,9 @@ func newMCPCmd() *cobra.Command {
 	var flags mcpFlags
 
 	cmd := &cobra.Command{
-		Use:   "mcp <agent-file>",
+		Use:   "mcp <agent-file>|<registry-ref>",
 		Short: "Start an agent as an MCP (Model Context Protocol) server",
-		Long:  `Start an stdio MCP server that exposes the agent via the Model Context Protocol`,
+		Long:  "Start an stdio MCP server that exposes the agent via the Model Context Protocol",
 		Example: `  cagent mcp ./agent.yaml
   cagent mcp ./team.yaml
   cagent mcp agentcatalog/pirate`,
@@ -47,5 +48,10 @@ func (f *mcpFlags) runMCPCommand(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	return mcp.StartMCPServer(ctx, out, args[0], f.runConfig)
+	agentFilename, err := agentfile.Resolve(ctx, out, args[0])
+	if err != nil {
+		return err
+	}
+
+	return mcp.StartMCPServer(ctx, out, agentFilename, f.runConfig)
 }
