+++ cmd/root/version.go
@@ -1,10 +1,9 @@
 package root
 
 import (
-	"fmt"
-
 	"github.com/spf13/cobra"
 
+	"github.com/docker/cagent/pkg/cli"
 	"github.com/docker/cagent/pkg/telemetry"
 	"github.com/docker/cagent/pkg/version"
 )
@@ -13,7 +12,7 @@ func newVersionCmd() *cobra.Command {
 	return &cobra.Command{
 		Use:   "version",
 		Short: "Print the version information",
-		Long:  `Display the version, build time, and commit hash`,
+		Long:  "Display the version and commit hash",
 		Args:  cobra.NoArgs,
 		Run:   runVersionCommand,
 	}
@@ -22,7 +21,7 @@ func newVersionCmd() *cobra.Command {
 func runVersionCommand(cmd *cobra.Command, args []string) {
 	telemetry.TrackCommand("version", args)
 
-	out := cmd.OutOrStdout()
-	fmt.Fprintf(out, "cagent version %s\n", version.Version)
-	fmt.Fprintf(out, "Commit: %s\n", version.Commit)
+	out := cli.NewPrinter(cmd.OutOrStdout())
+	out.Printf("cagent version %s\n", version.Version)
+	out.Printf("Commit: %s\n", version.Commit)
 }
